Use the `propertied` library to remove dependency on `kafka.client` (#20)
There was an implicit dependency on `kafka.client` through
`kafka.config`. That dependency was due to the need for a function to
convert a Clojure map to Java `Properties` object, which can be done
with the [`propertied`](https://github.com/michaelklishin/propertied)
library instead.

There's also a dependency on `clojure.tools.logging` that was not
specified in the `project.clj`.

Also remove local repo, because why not.