Simplify the serdes API, removing multimethod
This changeset removes the old `serde` multimethod, which presented a
significant opportunity for spooky action at a distance when users decided that
they needed to specialize the meaning of one of the "default" serdes.

The new API provides a default mapping, and lets users bring their own as they
need too instead of implicitly inviting them to participate in a global registry
for which no coherence could be provided.

It also simplifies the `resolve` function, which was previously special cased to
support parameterizing avro keys and values with their schemas.