[JACK-7] Re-implement the Avro Serde
The new API factors out the "registry" of Schema types to coercion
factories, making it possible for users to bring their own logical
type handlers when they instantiate serdes, just as #101 / JACK-13
makes it possible for users to bring their own schema registry
configurations.

This patch also rewrites much of the `jackdaw.avro.serdes` namespace,
trying to make sure that the coercion type stack is constructed only
when the serde is initialized. This should significantly optimize
serialization and deserialization, as previously the entire coercion
stack was rebuilt for every serialized or deserialized record.