Clean up serdes API.
This removes the `jackdaw.serdes` namespace and cleans up the underlying
serdes. The serde constructor was not really doing much for us and really
belonged in the app. This also renames top level functions in the
`jackdaw.serdes.avro` and `jackdaw.serdes.json` namespaces to match what is
done in `jackdaw.serdes.edn` and deletes some lesser used/lesser tested
serdes.

To create a serde, the usage is now uniform:

```
(jackdaw.serdes.<SERDE_TYPE>/serde <ARGS>)
```