Select matching record from union for coercion during deserialisation.
During the coercion phase of avro deserialisation, just assuming that a
given `GenericData$Record` is the correct one purely on the basis that
it is a `GenericData$Record` isn't enough if the record in question is
one of several in a union type - the result is that the first record
from the union every time.

So, change the logic so that the given
`Record` only is considered to "match" if the schema itself matches the
schema attached to the `Record` by avro itself.

Signed-off-by: Alistair Dutton <kelveden@gmail.com>