Add record name metadata to deserialized avro records
When an Avro record is deserialized to a Clojure map, add the record's
name and fully-qualified name to the map as metadata.
These record names are potentially useful info that was being discarded.

My personal use case for this metadata is to make it easier for
consuming code to deal with unions of records.

Signed-off-by: Sam Brauer <sam.brauer@gmail.com>