Our app, `loan-application-etl` quietly stopped processing messages over the weekend. It was still alive, but the last thing in the logs was:
```
WARN  org.apache.kafka.clients.consumer.internals.AbstractCoordinator: [Consumer clientId=consumer-1, groupId=loan-application-etl-1] This member will leave the group because consumer poll timeout has expired. This means the time between subsequent calls to poll() was longer than the configured max.poll.interval.ms, which typically implies that the poll loop is spending too much time processing messages. You can address this either by increasing max.poll.interval.ms or by reducing the maximum size of batches returned in poll() with max.poll.records.
```

I think we can increase the poll timeout as a partial solution, but it
would be better to properly handle this issue. If we can call the
version of `subscribe` that includes the `ConsumerRebalanceListener`,
then we can have an instance of it that simply throws an exception when
this happens, then allow marathon to restart the app.